requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.9.1
click==6.7
Django==1.10.6
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
mysqlclient==1.3.7
packaging==16.8
pi==0.1.2
pycparser==2.17
pyparsing==2.2.0
six==1.10.0
SQLAlchemy==1.1.6
virtualenv==15.1.0
Werkzeug==0.11.15
python-dateutil==2.6.0
